refactor(api): nest connections under gateway/ and tidy review findings
Move core/connections -> core/gateway/connections and
dbs/postgres/connections -> dbs/postgres/gateway/connections to match the
design docs, and rewrite all importers. Remove the dead
dbs/postgres/tools/{dao,dbes} leftovers. Use Field(default_factory=dict)
for the connection_data default, map ProviderNotFoundError to 404 in the
tools router, document get_provider, and assert 204 on the connection
delete acceptance tests (OSS + EE).
