+++ api/ee/tests/pytest/acceptance/tools/test_tools_connections.py
@@ -151,4 +151,5 @@ class TestToolsConnectionsLifecycle:
         assert revoke.status_code == 200, revoke.text
         assert revoke.json()["connection"]["flags"]["is_valid"] is False
 
-        connections_api("DELETE", f"/tools/connections/{connection_id}")
+        delete = connections_api("DELETE", f"/tools/connections/{connection_id}")
+        assert delete.status_code == 204, delete.text
