+++ api/oss/src/core/gateway/connections/exceptions.py
