+++ api/oss/src/core/gateway/connections/providers/composio/__init__.py
@@ -1,6 +1,6 @@
 # Avoid importing adapter here to prevent SDK dependency issues in standalone scripts.
 # Import directly when needed:
-#   from oss.src.core.connections.providers.composio.adapter import (
+#   from oss.src.core.gateway.connections.providers.composio.adapter import (
 #       ComposioConnectionsAdapter,
 #   )
 
@@ -12,7 +12,7 @@ __all__ = [
 def __getattr__(name):
     """Lazy import to avoid SDK dependency on module import."""
     if name == "ComposioConnectionsAdapter":
-        from oss.src.core.connections.providers.composio.adapter import (
+        from oss.src.core.gateway.connections.providers.composio.adapter import (
             ComposioConnectionsAdapter,
         )
