+++ api/oss/src/core/gateway/connections/utils.py
